валидация формы, обнуление состояния при переходе
