+++ src/components/Footer/Footer.js
@@ -1,4 +1,4 @@
-import './Footer.css'
+import './Footer.css';
 
 function Footer() {
   return (
