+++ src/components/Preloader/Preloader.js
@@ -4,7 +4,7 @@ function Preloader({ children = [] }, props) {
   return (
     <>
       <div className={props.preloader ? 'preloader preloader_open': 'preloader' }>
-        <div class="preloader__element">
+        <div className="preloader__element">
         </div>
       </div>
       {children}
