+++ src/components/SearchForm/SearchForm.css
@@ -1,3 +1,17 @@
+/* valid */
+.search-form__valid {
+  opacity: 0;
+  position: absolute;
+}
+
+.search-form__valid_active {
+  position: relative;
+  opacity: 1;
+}
+
+
+
+
 .search-form {
   margin-bottom: 80px;
   padding: 80px 70px 0;
